pyproject.toml
[tool.poetry]
name = "SysY TestGen"
version = "0.1.0"
description = "Generate SysY test file"
authors = ["Daniel <danielw10001@gmail.com>"]
license = "GPLv3"

[[tool.poetry.source]]
name = "tuna"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
default = true

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
